cleanup old functions
